Project 05: Computer Assembly
It has been a long journey, from nand to this final stage; put it all
together. First, we create the Memory component that should act as our
RAM with IO memory mapped to different addresses, the devices include
the Screen, Keyboard and RAM16K. Second, we assemble the CPU, activate
right circuitry as per the instruction input, from choosing ALU
arguments to choosing where the result is loaded and others, the CPU
does it. Thirdly, create the Computer Chip, this owns the ROM32K memory
chip where devices reside, Memory where program data is and the actual
CPU which connects it together bringing everything to life.

It is amazing that everything was built from nand gates.