Added Lexer for Jack
Jack will be my first type system to ever work with and I'll follow
rust's and c++'s features, shallowly. As per now, the plan is a very
simple type system with generics that are merely copy pasting type
arguments in where they occur, and a miniature version of the borrow
checker without lifetimes to simplify the work. With a function lexer
out of the way, next will be a parser. Also, they all fail on first
error, no blah, blah, blah.