Semantic analysis integrated within the parser
The original plan to make semantic analysis a tangible pass could not be
met given how low its impact would be, most of its work is doable by the
parser. Therefore, using context analysis, we were able to track and
allow/deny usage of tokens and expressions. As for resolution, we've
left that for the type system, that and type resolution go together.