Reduced bloat from VM
There is no need for a dedicated register to store THIS and THAT, as
they inflated the setup/teardown on calling functions, this results to
push 5 states just to call a getter method; RETURN_ADDRESS, LOCAL,
ARGUMENT, THIS and THAT. By assuming this will be the first argument, we
get to reuse ARGUMENT state effectively, the only offset needed is the
member offset to be accessed. But sometimes we need to access some
member of the value on top of the stack, simply RAM[STACK_TOP+OFFSET],
this has no need for a function call, just have the offset and the
object on top of the stack, and wala, getters and setters with little
overhead. This has reduced state to just 3; RETURN_ADDRESS, LOCAL and
ARGUMENT. the new instructions are `push/pop this I` (acting on ARG[0]
as this) and `push/pop member I` (acting on stack_top as this).