Test programs for new instructions
Since jack is an OOP language, making the VM strongly support OOP
semantics ease a big portion of Jack development, with that, compact
instructions have been added from `push/pop this/member I` to `push
FUNC` and `call NARGS`. These will guide in the creation of Jack
effectively. As for jack, the parser is now set and ready to rock, we'll
start with the most basic type system and evolve slowly, therefore,
generics will raise errors at first as they will not be implemented, but
later will be.